Fixed the 500 error when getting all rooms
Models/index: made the associations between User and Room easier to read. Added "player" alias to the many-to-many between User and Room
roomRoutes: changed casing on some aliases. Added the "player" alias to finding Room by ID
userRoutes: Identified the cause of the 500 error when getting a user by id